More progress on environment info, pita with dotnetcore shenanigans :(
