+++ source/Calamari/Integration/FileSystem/FileSizeExtensions.cs
@@ -2,23 +2,58 @@ namespace Calamari.Integration.FileSystem
 {
     public static class FileSizeExtensions
     {
-        private const long Kilobyte = 1024;
-        private const long Megabyte = 1024 * Kilobyte;
-        private const long Gigabyte = 1024 * Megabyte;
-        private const long Terabyte = 1024 * Gigabyte;
-
         public static string ToFileSizeString(this long bytes)
         {
             return ToFileSizeString(bytes <= 0 ? 0 : (ulong)bytes);
         }
 
-        public static string ToFileSizeString(this ulong bytes)
+        // Returns the human-readable file size for an arbitrary, 64-bit file size.
+        // The default format is "0.### XB", e.g. "4.2 KB" or "1.434 GB".
+        public static string ToFileSizeString(this ulong i)
         {
-            if (bytes > Terabyte) return (bytes / Terabyte).ToString("0 TB");
-            if (bytes > Gigabyte) return (bytes / Gigabyte).ToString("0 GB");
-            if (bytes > Megabyte) return (bytes / Megabyte).ToString("0 MB");
-            if (bytes > Kilobyte) return (bytes / Kilobyte).ToString("0 KB");
-            return bytes + " bytes";
+            // Determine the suffix and readable value.
+            string suffix;
+            double readable;
+            if (i >= 0x1000000000000000) // Exabyte
+            {
+                suffix = "EB";
+                readable = (i >> 50);
+            }
+            else if (i >= 0x4000000000000) // Petabyte
+            {
+                suffix = "PB";
+                readable = (i >> 40);
+            }
+            else if (i >= 0x10000000000) // Terabyte
+            {
+                suffix = "TB";
+                readable = (i >> 30);
+            }
+            else if (i >= 0x40000000) // Gigabyte
+            {
+                suffix = "GB";
+                readable = (i >> 20);
+            }
+            else if (i >= 0x100000) // Megabyte
+            {
+                suffix = "MB";
+                readable = (i >> 10);
+            }
+            else if (i >= 0x400) // Kilobyte
+            {
+                suffix = "KB";
+                readable = i;
+            }
+            else
+            {
+                return i.ToString("0 B"); // Byte
+            }
+
+            // Divide by 1024 to get fractional value.
+            readable = (readable / 1024);
+
+            // Return formatted number with suffix.
+            return readable.ToString("0.### ") + suffix;
         }
     }
 }
