+++ source/Calamari/Integration/Scripting/FSharp/Bootstrap.fsi
@@ -6,6 +6,7 @@ val private getEnvironmentVariable : name:string -> string option
 val private getCustomProxy : proxyHost:string -> System.Net.IWebProxy
 val private getCustomCredentials : proxyUserName:string -> System.Net.NetworkCredential
 val private decryptString : encrypted:string -> iv:string -> string
+val private logEnvironmentInformation : unit -> unit
 val tryFindVariable : name:string -> string option
 val findVariable : name:string -> string
 val findVariableOrDefault : defaultValue:string -> name:string -> string
