+++ source/Calamari/Integration/Scripting/FSharp/FSharpBootstrapper.cs
@@ -4,6 +4,8 @@ using System.Text;
 using Calamari.Commands.Support;
 using Calamari.Integration.Processes;
 using Calamari.Util;
+using Calamari.Util.Environments;
+using System.Reflection; /* Don't remove this, it overrides the Calamari.Util.CrossPlatformExtensions */
 
 namespace Calamari.Integration.Scripting.FSharp
 {
@@ -64,7 +66,8 @@ namespace Calamari.Integration.Scripting.FSharp
             var configurationFile = Path.Combine(workingDirectory, "Configure." + Guid.NewGuid().ToString().Substring(10) + ".fsx");
 
             var builder = new StringBuilder(BootstrapScriptTemplate);
-            builder.Replace("{{VariableDeclarations}}", WritePatternMatching(variables));
+            builder.Replace("(*{{VariableDeclarations}}*)", WritePatternMatching(variables));
+            builder.Replace("(*{{LogEnvironmentInformation}}*)", LogEnvironmentInformation());
 
             using (var file = new FileStream(configurationFile, FileMode.CreateNew, FileAccess.Write))
             using (var writer = new StreamWriter(file, Encoding.UTF8))
@@ -76,7 +79,19 @@ namespace Calamari.Integration.Scripting.FSharp
             File.SetAttributes(configurationFile, FileAttributes.Hidden);
             return configurationFile;
         }
-            
+
+        static string LogEnvironmentInformation()
+        {
+            var output = new StringBuilder();
+
+            foreach (var envInfo in EnvironmentHelper.SafelyGetEnvironmentInformation())
+            {
+                output.AppendLine($"printfn \"{envInfo}\"");
+            }
+
+            return output.ToString();
+        }
+
         static string WritePatternMatching(CalamariVariableDictionary variables)
         {
             var builder = new StringBuilder();
