+++ source/Calamari/project.json
@@ -78,6 +78,7 @@
         "System.Reflection.Extensions": "4.0.1",
         "System.Reflection.TypeExtensions": "4.1.0",
         "System.Runtime": "4.1.0",
+        "System.Runtime.Extensions": "4.1.0",
         "System.Runtime.InteropServices.RuntimeInformation": "4.0.0",
         "System.Runtime.Serialization.Formatters": "4.0.0-rc3-24212-01",
         "System.Security.Cryptography.Primitives": "4.0.0",
